Вынес shopping card в отдельный глобальный компонент
