+++ frontend/src/components/shopping-list-card/ShoppingListContent.tsx
