+++ frontend/src/components/shopping-list-card/ShoppingListHeader.tsx
