+++ frontend/src/components/shopping-list-card/ShoppingListItem.tsx
