+++ frontend/src/components/shopping-list-card/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ShoppingListCard';
